Change #toDouble():Double -> #toNumber():Number and add #createNumber(long):T
Just because the primary case is to process JSON that doesn't mean we can assume everyone wants to treat all numbers as doubles. This changes the API so that the adapter will not have to convert numbers (toDouble -> toNumber) and functions like length will not have to return their values as doubles.

Functions like abs, ceil and floor still convert to doubles (the alternative for abs would be to check the input type, and it's doubtful that it's worth it); sort will convert non-doubles to doubles before comparing them.

+++ jmespath-core/src/main/java/io/burt/jmespath/function/LengthFunction.java
@@ -17,9 +17,9 @@ public class LengthFunction extends JmesPathFunction {
     } else {
       T subject = argument.value();
       if (adapter.isString(subject)) {
-        return adapter.createNumber(adapter.toString(subject).length());
+        return adapter.createNumber((long) adapter.toString(subject).length());
       } else if (adapter.isArray(subject) || adapter.isObject(subject)) {
-        return adapter.createNumber(adapter.toList(subject).size());
+        return adapter.createNumber((long) adapter.toList(subject).size());
       } else {
         throw new ArgumentTypeException(name(), "string, array or object", adapter.typeOf(subject).toString());
       }
